requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
phonenumbers==8.13.35
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
setuptools==69.5.1
sqlparse==0.5.0
uritemplate==4.1.1
